extract solana-precompiles crate
